itemfactory will be instanceable powerups are active by default acceleratation/retardation are ints not floats item is observable so that it can notify when it should be removed unit has some new methods for PUs but not final yet

+++ project/src/model/powerup/IPowerUp.java
@@ -12,7 +12,7 @@ public interface IPowerUp {
 
     public boolean isActive();
 
-    public float getAcceleration();
+    public int getAcceleration();
 
     public int getHitPoints();
 
@@ -22,7 +22,7 @@ public interface IPowerUp {
 
     public float getMaxSpeed();
 
-    public float getRetardation();
+    public int getRetardation();
 
     public int getSteerAngle();
